requirements.txt
langchain
langchain_community
langchain_ollama
ollama
opencc-python-reimplemented
sentence-transformers
numpy
pymilvus